+++ app/include/tab/setting_tab.hpp
@@ -36,6 +36,7 @@ private:
     BRLS_BIND(brls::DetailCell, btnUser, "setting/user");
     BRLS_BIND(brls::DetailCell, btnServer, "setting/server");
     BRLS_BIND(brls::BooleanCell, btnHWDEC, "setting/video/hwdec");
+    BRLS_BIND(brls::BooleanCell, btnQuality, "setting/video/low_quality");
     BRLS_BIND(brls::BooleanCell, btnDirectPlay, "setting/video/directplay");
     BRLS_BIND(brls::SelectorCell, selectorCodec, "setting/video/codec");
     BRLS_BIND(brls::SelectorCell, selectorQuality, "setting/video/quality");